Fix number strings being converted to numbers
Fixes #78. This also fixes numbers larger than
Number.MAX_SAFE_INTEGER represented as strings being rounded
when the yaml is loaded and serialized again.

Causes a slight regression with the styling of string values.
Specifically, any strings that are modified in the JSON are quoted
in the resulting yaml, even if they weren't in the original.